The first diff is for the AddUserForm.tsx file. In this file, the author has removed some headers from the code. This was likely done in order to simplify the code or remove unnecessary information.
The second diff is for the api/users.ts file. In this file, the author has changed the way that the delete function works. Previously, it was using the body of the request to determine the id of the user to delete. However, it has been changed to use the query string instead. This is likely a more efficient way of doing this operation.

The third diff is for the users.tsx file. In this file, the author has added a new function to handle deleting users. This function makes a request to the backend to delete the user with the specified id. Once the user is deleted, the page is refreshed.